sema: move `Attribute` to separate file
